refactor(core): extract BasePeerRunner for DreamerRunner first slice (PRI-302)
Extract BasePeerRunner<TContext, TOutput> abstract class from duplicated
peer runner pipeline code. Migrate DreamerRunner as proof of concept.

New files:
- runner/base-peer-runner.ts: abstract base class with shared lease →
  buildContext → invoke → poll → fetch → validate → succeed pipeline
- runner/peer-runner-types.ts: shared types (PeerRunnerOptions,
  PeerRunnerDeps, PeerRunnerConfig, PeerRunnerResult, etc.)

Modified:
- dreamer-runner.ts: extends BasePeerRunner, only Dreamer-specific logic
  remains (~200 lines vs ~830 lines original)
- index.ts: exports new base class and shared types

Design decisions (confirmed via grilling + review):
- succeedTask() is abstract (DiagnosticianCommitter vs PIArtifactStore)
- validateOutput() is abstract (delegates to injected validator)
- retryOrFail() in base class with try/catch robustness pattern
- postFetchTransform() hook for lineage field re-injection
- checkLineageIntegrity() optional hook for lineage strip contract
- PeerRunnerConfig passed to constructor (runnerName, expectedTaskKind, etc.)

Scope: Phase A only. Remaining 7 runners in follow-up PRs.

Verification:
- principles-core build: PASS
- 744 tests (16 files): PASS
- architecture-regression (440 tests): PASS
- DiagnosticianRunner tests (34 tests): PASS (untouched)
- openclaw-plugin build: PASS
- lint: PASS
