+++ packages/principles-core/src/runtime-v2/runner/base-peer-runner.ts
@@ -1,0 +1,632 @@
+/**
+ * BasePeerRunner — Abstract base class for all peer runners (PRI-302).
+ *
+ * Extracts the shared lease → buildContext → invoke → poll → fetch →
+ * validate → succeed/fail pipeline from 8 duplicated runner implementations.
+ *
+ * Subclasses implement:
+ *   - permanentErrorCategories (abstract getter)
+ *   - buildContext() — runner-specific context assembly
+ *   - invokeRuntime() — runner-specific LLM invocation
+ *   - validateOutput() — runner-specific output validation
+ *   - succeedTask() — runner-specific artifact commit + task success
+ *
+ * Optional hooks:
+ *   - emitSuccessTelemetry() — runner-specific success telemetry (default: no-op)
+ *   - checkLineageIntegrity() — lineage strip contract check (default: no-op)
+ *
+ * @see docs/adr/0003-peer-agent-state-machine-orchestration.md
+ * @see Linear PRI-302
+ */
+
+import type { RuntimeStateManager } from '../store/runtime-state-manager.js';
+import type {
+  PDRuntimeAdapter,
+  RunHandle,
+  RunStatus,
+} from '../runtime-protocol.js';
+import type { StoreEventEmitter } from '../store/event-emitter.js';
+import type { PIArtifactStore } from '../internalization/pi-artifact.js';
+import type { TaskRecord } from '../task-status.js';
+import { PDRuntimeError, type PDErrorCategory } from '../error-categories.js';
+import type { TelemetryEvent } from '../../telemetry-event.js';
+import { RunnerPhase } from './runner-phase.js';
+import type {
+  PeerRunnerOptions,
+  ResolvedPeerRunnerOptions,
+  PeerRunnerDeps,
+  PeerRunnerConfig,
+  PeerRunnerResult,
+  PeerRunnerValidationResult,
+  FailureContext,
+} from './peer-runner-types.js';
+
+// ── Defaults ─────────────────────────────────────────────────────────────────
+
+const DEFAULT_PEER_RUNNER_OPTIONS: Readonly<Omit<ResolvedPeerRunnerOptions, 'owner' | 'runtimeKind'>> = {
+  pollIntervalMs: 5_000,
+  timeoutMs: 300_000,
+  defaultMaxAttempts: 3,
+  agentId: 'peer-runner',
+} as const;
+
+function resolvePeerRunnerOptions(
+  options: PeerRunnerOptions,
+  defaultAgentId: string,
+): ResolvedPeerRunnerOptions {
+  return {
+    pollIntervalMs: options.pollIntervalMs ?? DEFAULT_PEER_RUNNER_OPTIONS.pollIntervalMs,
+    timeoutMs: options.timeoutMs ?? DEFAULT_PEER_RUNNER_OPTIONS.timeoutMs,
+    defaultMaxAttempts: options.defaultMaxAttempts ?? DEFAULT_PEER_RUNNER_OPTIONS.defaultMaxAttempts,
+    owner: options.owner,
+    runtimeKind: options.runtimeKind,
+    agentId: options.agentId ?? defaultAgentId,
+  };
+}
+
+// ── BasePeerRunner ───────────────────────────────────────────────────────────
+
+/**
+ * Abstract base class for peer runners.
+ *
+ * TContext — the type returned by buildContext(). Must include contextHash.
+ * TOutput — the type of the validated LLM output.
+ */
+export abstract class BasePeerRunner<TContext extends { contextHash: string }, TOutput> {
+  // ── Shared dependencies (protected for subclass access) ──
+  protected readonly stateManager: RuntimeStateManager;
+  protected readonly runtimeAdapter: PDRuntimeAdapter;
+  protected readonly eventEmitter: StoreEventEmitter;
+  protected readonly artifactStore: PIArtifactStore;
+  protected readonly resolvedOptions: ResolvedPeerRunnerOptions;
+  protected readonly config: PeerRunnerConfig;
+  private phase: RunnerPhase = RunnerPhase.Idle;
+
+  constructor(
+    deps: PeerRunnerDeps,
+    options: PeerRunnerOptions,
+    config: PeerRunnerConfig,
+  ) {
+    this.stateManager = deps.stateManager;
+    this.runtimeAdapter = deps.runtimeAdapter;
+    this.eventEmitter = deps.eventEmitter;
+    this.artifactStore = deps.artifactStore;
+    this.config = config;
+    this.resolvedOptions = resolvePeerRunnerOptions(options, config.defaultAgentId);
+  }
+
+  // ── Observability ──────────────────────────────────────────────────────────
+
+  /** Current internal phase. For testing/observability only. */
+  get currentPhase(): RunnerPhase {
+    return this.phase;
+  }
+
+  // ── Abstract: subclass must implement ───────────────────────────────────────
+
+  /** Permanent error categories — runner's inherent property. */
+  abstract get permanentErrorCategories(): ReadonlySet<PDErrorCategory>;
+
+  /** Build runner-specific context from task and predecessor outputs. */
+  abstract buildContext(taskId: string): Promise<TContext>;
+
+  /** Invoke the runtime with runner-specific prompt builder. */
+  abstract invokeRuntime(taskId: string, context: TContext): Promise<RunHandle>;
+
+  /** Validate the LLM output. */
+  abstract validateOutput(output: TOutput, taskId: string, context: TContext): Promise<PeerRunnerValidationResult>;
+
+  /** Commit artifact + mark task succeeded. Runner-specific commit strategy. */
+  abstract succeedTask(
+    taskId: string,
+    runId: string,
+    output: TOutput,
+    task: TaskRecord,
+    contextHash: string,
+    context: TContext,
+  ): Promise<PeerRunnerResult<TOutput>>;
+
+  // ── Optional hooks (default no-op) ─────────────────────────────────────────
+
+  /** Emit runner-specific success telemetry. Called after validation passes. */
+  // eslint-disable-next-line @typescript-eslint/class-methods-use-this
+  protected emitSuccessTelemetry(_taskId: string, _output: TOutput, _context: TContext): void {
+    // default no-op
+  }
+
+  /** Check lineage strip contract. Called after fetchAndParseOutput. */
+  // eslint-disable-next-line @typescript-eslint/class-methods-use-this
+  protected checkLineageIntegrity(_taskId: string, _output: TOutput): void {
+    // default no-op
+  }
+
+  /**
+   * Transform output after fetch, before validation.
+   * Used by runners that need to re-inject lineage fields stripped by the adapter.
+   * Called after fetchAndParseOutput, before validateOutput.
+   */
+  // eslint-disable-next-line @typescript-eslint/class-methods-use-this
+  protected postFetchTransform(_taskId: string, _output: TOutput): void {
+    // default no-op
+  }
+
+  // ── Template method: full pipeline ─────────────────────────────────────────
+
+  /**
+   * Execute the full peer runner lifecycle for a task.
+   *
+   * Pipeline: lease → resolveRunId → buildContext → invoke → poll →
+   *           fetch → validate → succeedTask (abstract)
+   *
+   * Each invocation is independent — no mutable state between run() calls.
+   */
+  async run(taskId: string): Promise<PeerRunnerResult<TOutput>> {
+    this.phase = RunnerPhase.Idle;
+
+    // 1. Acquire lease — isolated try/catch so lease_conflict never uses synthetic TaskRecord
+    let leasedTask: TaskRecord;
+    try {
+      leasedTask = await this.stateManager.acquireLease({
+        taskId,
+        owner: this.resolvedOptions.owner,
+        runtimeKind: this.resolvedOptions.runtimeKind,
+      });
+    } catch (error) {
+      return await this.handleLeaseOrPhaseError(taskId, error);
+    }
+
+    // 1b. Validate task kind
+    if (leasedTask.taskKind !== this.config.expectedTaskKind) {
+      this.emitEvent('wrong_task_kind', taskId, {
+        expectedKind: this.config.expectedTaskKind,
+        actualKind: leasedTask.taskKind,
+      });
+      await this.stateManager.markTaskFailed(taskId, 'input_invalid');
+      return {
+        status: 'failed',
+        taskId,
+        errorCategory: 'input_invalid',
+        failureReason: `Task kind must be '${this.config.expectedTaskKind}', got '${leasedTask.taskKind}'`,
+        attemptCount: leasedTask.attemptCount,
+      };
+    }
+
+    this.emitEvent('task_leased', taskId, {
+      taskKind: this.config.expectedTaskKind,
+      attemptCount: leasedTask.attemptCount,
+    });
+
+    // All subsequent errors use the real leasedTask — no synthetic TaskRecord allowed
+    try {
+      // 2. Resolve store runId
+      const storeRunId = await this.resolveStoreRunId(taskId);
+
+      // 3. Build context
+      this.phase = RunnerPhase.BuildingContext;
+      const context = await this.buildContext(taskId);
+      this.emitEvent('context_built', taskId, { contextHash: context.contextHash });
+
+      // 4. Invoke runtime
+      this.phase = RunnerPhase.Invoking;
+      const runHandle = await this.invokeRuntime(taskId, context);
+      this.emitEvent('run_started', taskId, {
+        runtimeKind: this.resolvedOptions.runtimeKind,
+      });
+
+      // 5. Poll until terminal
+      this.phase = RunnerPhase.Polling;
+      const finalStatus = await this.pollUntilTerminal(runHandle);
+
+      // 6. Handle non-success terminal states
+      if (finalStatus.status !== 'succeeded') {
+        return await this.handleRuntimeFailure(taskId, leasedTask, finalStatus);
+      }
+
+      // 7. Fetch output
+      this.phase = RunnerPhase.FetchingOutput;
+      const output = await this.fetchAndParseOutput(runHandle.runId, taskId);
+
+      // 7b. Post-fetch transform (re-inject lineage fields, etc.)
+      this.postFetchTransform(taskId, output);
+
+      // 7c. Check lineage integrity (optional hook)
+      this.checkLineageIntegrity(taskId, output);
+
+      // 8. Validate
+      this.phase = RunnerPhase.Validating;
+      const validationResult = await this.validateOutput(output, taskId, context);
+      if (!validationResult.valid) {
+        return await this.handleValidationError({
+          taskId,
+          task: leasedTask,
+          errors: validationResult.errors,
+          errorCategory: validationResult.errorCategory,
+        });
+      }
+
+      this.emitEvent('output_validated', taskId, {});
+
+      // 8b. Emit success telemetry (optional hook)
+      this.emitSuccessTelemetry(taskId, output, context);
+
+      // 9. Succeed task (abstract — subclass implements commit strategy)
+      return await this.succeedTask(taskId, storeRunId, output, leasedTask, context.contextHash, context);
+    } catch (error) {
+      // Post-lease errors use retryOrFail with the real leasedTask
+      return await this.handlePostLeaseError(taskId, leasedTask, error);
+    }
+  }
+
+  // ── Telemetry ──────────────────────────────────────────────────────────────
+
+  /**
+   * Emit a telemetry event with the runner's name prefix.
+   * Event type: `{runnerName}_{eventType}`
+   */
+  protected emitEvent(
+    eventType: string,
+    taskId: string,
+    payload: Record<string, unknown>,
+  ): void {
+    this.eventEmitter.emitTelemetry({
+      eventType: `${this.config.runnerName}_${eventType}` as TelemetryEvent['eventType'],
+      traceId: taskId,
+      timestamp: new Date().toISOString(),
+      sessionId: this.resolvedOptions.owner,
+      agentId: this.resolvedOptions.agentId,
+      payload,
+    });
+  }
+
+  // ── Polling ────────────────────────────────────────────────────────────────
+
+  private async pollUntilTerminal(runHandle: RunHandle): Promise<RunStatus> {
+    const deadline = Date.now() + this.resolvedOptions.timeoutMs;
+    const terminalStatuses: readonly string[] = ['succeeded', 'failed', 'timed_out', 'cancelled'];
+
+    while (Date.now() < deadline) {
+      const status = await this.runtimeAdapter.pollRun(runHandle.runId);
+      if (terminalStatuses.includes(status.status)) {
+        return status;
+      }
+      await this.sleep(this.resolvedOptions.pollIntervalMs);
+    }
+
+    // Timeout — cancel gracefully
+    let cancelFailed = false;
+    try {
+      await this.runtimeAdapter.cancelRun(runHandle.runId);
+    } catch (cancelErr) {
+      cancelFailed = true;
+      this.emitEvent('cancel_run_failed', runHandle.runId, {
+        runId: runHandle.runId,
+        errorMessage: cancelErr instanceof Error ? cancelErr.message : String(cancelErr),
+      });
+    }
+    const cancelNote = cancelFailed ? ' (cancelRun also failed)' : '';
+    throw new PDRuntimeError('timeout', `Run ${runHandle.runId} timed out after ${this.resolvedOptions.timeoutMs}ms${cancelNote}`);
+  }
+
+  // ── Output fetching ────────────────────────────────────────────────────────
+
+  private async fetchAndParseOutput(runId: string, taskId: string): Promise<TOutput> {
+    let result: Awaited<ReturnType<PDRuntimeAdapter['fetchOutput']>>;
+    try {
+      result = await this.runtimeAdapter.fetchOutput(runId);
+    } catch (fetchErr) {
+      this.emitEvent('output_extraction_failed', taskId, {
+        runId,
+        stage: 'fetchOutput',
+        errorMessage: fetchErr instanceof Error ? fetchErr.message : String(fetchErr),
+      });
+      throw fetchErr;
+    }
+
+    if (!result || !result.payload) {
+      this.emitEvent('output_extraction_failed', taskId, {
+        runId,
+        stage: 'payload_missing',
+        errorMessage: `No output available for run ${runId}`,
+      });
+      throw new PDRuntimeError('output_invalid', `No output available for run ${runId}`);
+    }
+
+    const payload = result.payload as Record<string, unknown>;
+    if (typeof payload !== 'object' || payload === null) {
+      this.emitEvent('output_extraction_failed', taskId, {
+        runId,
+        stage: 'payload_not_object',
+        errorMessage: `Output payload is not an object for run ${runId}`,
+      });
+      throw new PDRuntimeError('output_invalid', `Output payload is not an object for run ${runId}`);
+    }
+
+    return result.payload as TOutput;
+  }
+
+  // ── Lineage resolution ─────────────────────────────────────────────────────
+
+  /** Resolve artifact IDs from predecessor tasks for lineage tracking. */
+  protected async resolveLineageArtifactIds(taskId: string): Promise<{ ids: string[]; hasRejected: boolean }> {
+    const { hydratePITaskRecord } = await import('../internalization/pitask-metadata.js');
+    const task = await this.stateManager.getTask(taskId);
+    if (!task) return { ids: [], hasRejected: false };
+
+    const piTask = hydratePITaskRecord(task);
+    const deps = piTask?.dependencyTaskIds ?? [];
+    if (deps.length === 0) return { ids: [], hasRejected: false };
+
+    const lineageIds: string[] = [];
+    let hasRejected = false;
+    const results = await Promise.allSettled(
+      deps.map((depId) => this.artifactStore.listBySourceTaskId(depId)),
+    );
+    for (const result of results) {
+      if (result.status === 'fulfilled') {
+        for (const artifact of result.value) {
+          lineageIds.push(artifact.artifactId);
+        }
+      } else {
+        hasRejected = true;
+      }
+    }
+    return { ids: lineageIds, hasRejected };
+  }
+
+  // ── Store helpers ──────────────────────────────────────────────────────────
+
+  private async resolveStoreRunId(taskId: string): Promise<string> {
+    const runs = await this.stateManager.getRunsByTask(taskId);
+    const latestRun = runs[runs.length - 1];
+    if (!latestRun) {
+      throw new PDRuntimeError('execution_failed', `No run records found for task ${taskId} after lease acquisition`);
+    }
+    return latestRun.runId;
+  }
+
+  /** Compute a deterministic hash from context references (observability only). */
+  protected static hashContextRefs(refs: readonly string[]): string {
+    if (refs.length === 0) return 'empty';
+    const str = refs.join('|');
+    let hash = 0;
+    for (let i = 0; i < str.length; i++) {
+      hash = (Math.imul(31, hash) + str.charCodeAt(i)) | 0;
+    }
+    return `ctx-${Math.abs(hash).toString(16)}`;
+  }
+
+  // ── Error handling ─────────────────────────────────────────────────────────
+
+  private async handleLeaseOrPhaseError(
+    taskId: string,
+    error: unknown,
+  ): Promise<PeerRunnerResult<TOutput>> {
+    const classified = this.classifyError(error);
+
+    // lease_conflict is concurrent-access conflict, not a state change.
+    // No mutation methods (markTaskFailed/markTaskRetryWait) are called.
+    if (classified.category === 'lease_conflict') {
+      this.emitEvent('run_failed', taskId, {
+        errorCategory: 'lease_conflict',
+        errorMessage: classified.message,
+      });
+      return {
+        status: 'failed',
+        taskId,
+        errorCategory: 'lease_conflict',
+        failureReason: classified.message,
+        attemptCount: 1,
+      };
+    }
+
+    // Non-lease errors (e.g., storage_unavailable before lease) must not
+    // use synthetic TaskRecord. Build one with real defaults from options.
+    this.emitEvent('run_failed', taskId, {
+      errorCategory: classified.category,
+      errorMessage: classified.message,
+    });
+
+    const task: TaskRecord = {
+      taskId,
+      taskKind: this.config.expectedTaskKind,
+      status: 'leased',
+      createdAt: new Date().toISOString(),
+      updatedAt: new Date().toISOString(),
+      attemptCount: 1,
+      maxAttempts: this.resolvedOptions.defaultMaxAttempts,
+    };
+    return this.retryOrFail({ taskId, task, errorCategory: classified.category, failureReason: classified.message });
+  }
+
+  private async handlePostLeaseError(
+    taskId: string,
+    task: TaskRecord,
+    error: unknown,
+  ): Promise<PeerRunnerResult<TOutput>> {
+    const classified = this.classifyError(error);
+
+    this.emitEvent('run_failed', taskId, {
+      errorCategory: classified.category,
+      errorMessage: classified.message,
+    });
+
+    return this.retryOrFail({ taskId, task, errorCategory: classified.category, failureReason: classified.message });
+  }
+
+  private async handleRuntimeFailure(
+    taskId: string,
+    task: TaskRecord,
+    runStatus: RunStatus,
+  ): Promise<PeerRunnerResult<TOutput>> {
+    const errorCategory = this.mapRunStatusToErrorCategory(runStatus.status, runStatus.reason);
+
+    this.emitEvent('run_failed', taskId, {
+      runStatus: runStatus.status,
+      errorCategory,
+    });
+
+    return this.retryOrFail({
+      taskId,
+      task,
+      errorCategory,
+      failureReason: `Runtime execution ended with status: ${runStatus.status}`,
+    });
+  }
+
+  protected async handleValidationError(ctx: {
+    taskId: string;
+    task: TaskRecord;
+    errors: readonly string[];
+    errorCategory?: PDErrorCategory;
+  }): Promise<PeerRunnerResult<TOutput>> {
+    const category = ctx.errorCategory ?? 'output_invalid';
+
+    this.emitEvent('output_invalid', ctx.taskId, {
+      errorCount: ctx.errors.length,
+      errorCategory: category,
+    });
+
+    return this.retryOrFail({
+      taskId: ctx.taskId,
+      task: ctx.task,
+      errorCategory: category,
+      failureReason: `Validation failed: ${ctx.errors.join('; ')}`,
+    });
+  }
+
+  // ── Retry/fail state machine ───────────────────────────────────────────────
+
+  /**
+   * Core retry-or-fail decision.
+   *
+   * Uses try/catch around markTaskFailed/markTaskRetryWait for robustness.
+   * If state manager operations fail, returns storage_unavailable instead of
+   * propagating the exception (ERR-002: graceful degradation with reason).
+   */
+  protected async retryOrFail(ctx: FailureContext): Promise<PeerRunnerResult<TOutput>> {
+    // Permanent errors — never retry
+    if (this.permanentErrorCategories.has(ctx.errorCategory)) {
+      try {
+        await this.stateManager.markTaskFailed(ctx.taskId, ctx.errorCategory);
+      } catch (stateErr) {
+        this.emitEvent('mark_failed_error', ctx.taskId, {
+          errorCategory: 'storage_unavailable',
+          attemptCount: ctx.task.attemptCount,
+          errorMessage: stateErr instanceof Error ? stateErr.message : String(stateErr),
+        });
+        return {
+          status: 'failed',
+          taskId: ctx.taskId,
+          errorCategory: 'storage_unavailable',
+          failureReason: `State manager error: ${ctx.failureReason}`,
+          attemptCount: ctx.task.attemptCount,
+        };
+      }
+      this.emitEvent('task_failed', ctx.taskId, {
+        errorCategory: ctx.errorCategory,
+        attemptCount: ctx.task.attemptCount,
+        failureReason: ctx.failureReason,
+      });
+      this.phase = RunnerPhase.Failed;
+      return {
+        status: 'failed',
+        taskId: ctx.taskId,
+        errorCategory: ctx.errorCategory,
+        failureReason: ctx.failureReason,
+        attemptCount: ctx.task.attemptCount,
+      };
+    }
+
+    // Retry policy check
+    const shouldRetry = this.stateManager.getRetryPolicy().shouldRetry(ctx.task);
+    if (shouldRetry) {
+      try {
+        await this.stateManager.markTaskRetryWait(ctx.taskId, ctx.errorCategory);
+      } catch (stateErr) {
+        this.emitEvent('mark_retry_error', ctx.taskId, {
+          errorCategory: 'storage_unavailable',
+          attemptCount: ctx.task.attemptCount,
+          errorMessage: stateErr instanceof Error ? stateErr.message : String(stateErr),
+        });
+        return {
+          status: 'failed',
+          taskId: ctx.taskId,
+          errorCategory: 'storage_unavailable',
+          failureReason: `State manager error: ${ctx.failureReason}`,
+          attemptCount: ctx.task.attemptCount,
+        };
+      }
+      this.emitEvent('task_retried', ctx.taskId, {
+        errorCategory: ctx.errorCategory,
+        attemptCount: ctx.task.attemptCount,
+      });
+      this.phase = RunnerPhase.RetryWaiting;
+      return {
+        status: 'retried',
+        taskId: ctx.taskId,
+        errorCategory: ctx.errorCategory,
+        failureReason: ctx.failureReason,
+        attemptCount: ctx.task.attemptCount,
+      };
+    }
+
+    // Max attempts exceeded
+    try {
+      await this.stateManager.markTaskFailed(ctx.taskId, 'max_attempts_exceeded');
+    } catch (stateErr) {
+      this.emitEvent('mark_failed_error', ctx.taskId, {
+        errorCategory: 'storage_unavailable',
+        attemptCount: ctx.task.attemptCount,
+        errorMessage: stateErr instanceof Error ? stateErr.message : String(stateErr),
+      });
+      return {
+        status: 'failed',
+        taskId: ctx.taskId,
+        errorCategory: 'storage_unavailable',
+        failureReason: `State manager error: ${ctx.failureReason}`,
+        attemptCount: ctx.task.attemptCount,
+      };
+    }
+    this.emitEvent('task_failed', ctx.taskId, {
+      errorCategory: 'max_attempts_exceeded',
+      attemptCount: ctx.task.attemptCount,
+      failureReason: `Max attempts exceeded: ${ctx.failureReason}`,
+    });
+    this.phase = RunnerPhase.Failed;
+    return {
+      status: 'failed',
+      taskId: ctx.taskId,
+      errorCategory: 'max_attempts_exceeded',
+      failureReason: `Max attempts exceeded: ${ctx.failureReason}`,
+      attemptCount: ctx.task.attemptCount,
+    };
+  }
+
+  // ── Error classification ──────────────────────────────────────────────────
+
+  // eslint-disable-next-line @typescript-eslint/class-methods-use-this
+  private classifyError(error: unknown): { category: PDErrorCategory; message: string } {
+    if (error instanceof PDRuntimeError) {
+      return { category: error.category, message: error.message };
+    }
+    if (error instanceof Error) {
+      return { category: 'execution_failed', message: error.message };
+    }
+    return { category: 'execution_failed', message: String(error) };
+  }
+
+  // eslint-disable-next-line @typescript-eslint/class-methods-use-this
+  private mapRunStatusToErrorCategory(status: string, _reason?: string): PDErrorCategory {
+    switch (status) {
+      case 'failed': return 'execution_failed';
+      case 'timed_out': return 'timeout';
+      case 'cancelled': return 'cancelled';
+      default: return 'execution_failed';
+    }
+  }
+
+  // eslint-disable-next-line @typescript-eslint/class-methods-use-this
+  private sleep(ms: number): Promise<void> {
+    return new Promise((resolve) => setTimeout(resolve, ms));
+  }
+}
