+++ packages/principles-core/src/runtime-v2/runner/peer-runner-types.ts
@@ -1,0 +1,128 @@
+/**
+ * Shared types for the BasePeerRunner abstract class (PRI-302).
+ *
+ * These types replace the per-runner duplicated interfaces:
+ *   - DreamerRunnerOptions / ScribeRunnerOptions / ...
+ *   - DreamerRunnerDeps / ScribeRunnerDeps / ...
+ *   - DreamerRunnerResult / ScribeRunnerResult / ...
+ *   - FailureContext / SucceedContext / ValidationErrorContext
+ *
+ * @see docs/adr/0003-peer-agent-state-machine-orchestration.md
+ */
+
+import type { PDErrorCategory } from '../error-categories.js';
+import type { RuntimeStateManager } from '../store/runtime-state-manager.js';
+import type { PDRuntimeAdapter } from '../runtime-protocol.js';
+import type { StoreEventEmitter } from '../store/event-emitter.js';
+import type { PIArtifactStore } from '../internalization/pi-artifact.js';
+import type { PeerRunnerKind } from '../internalization/peer-runner-contracts.js';
+import type { TaskRecord } from '../task-status.js';
+
+// ── Options ──────────────────────────────────────────────────────────────────
+
+/** Shared runner options. All peer runners accept this shape. */
+export interface PeerRunnerOptions {
+  readonly pollIntervalMs?: number;
+  readonly timeoutMs?: number;
+  readonly defaultMaxAttempts?: number;
+  readonly owner: string;
+  readonly runtimeKind: string;
+  readonly agentId?: string;
+}
+
+/** Resolved options with defaults applied. */
+export interface ResolvedPeerRunnerOptions {
+  readonly pollIntervalMs: number;
+  readonly timeoutMs: number;
+  readonly defaultMaxAttempts: number;
+  readonly owner: string;
+  readonly runtimeKind: string;
+  readonly agentId: string;
+}
+
+// ── Dependencies ─────────────────────────────────────────────────────────────
+
+/**
+ * Shared dependencies injected into all peer runners.
+ *
+ * Runners with additional dependencies (e.g. DiagnosticianRunner needs
+ * ContextAssembler + DiagnosticianCommitter) should extend this interface.
+ */
+export interface PeerRunnerDeps {
+  readonly stateManager: RuntimeStateManager;
+  readonly runtimeAdapter: PDRuntimeAdapter;
+  readonly eventEmitter: StoreEventEmitter;
+  readonly artifactStore: PIArtifactStore;
+}
+
+// ── Config ───────────────────────────────────────────────────────────────────
+
+/**
+ * Runner-specific configuration passed to the BasePeerRunner constructor.
+ * Each subclass provides its own config instance.
+ */
+export interface PeerRunnerConfig {
+  /** Runner name for telemetry event prefix (e.g. 'dreamer', 'scribe'). */
+  readonly runnerName: string;
+  /** Expected taskKind for lease validation (e.g. 'dreamer'). */
+  readonly expectedTaskKind: PeerRunnerKind;
+  /** Default agentId if not provided in options. */
+  readonly defaultAgentId: string;
+  /** ResultRef prefix (e.g. 'dreamer' → 'dreamer://runId'). */
+  readonly resultRefPrefix: string;
+}
+
+// ── Result ───────────────────────────────────────────────────────────────────
+
+/** Status values shared by all peer runner results. */
+export type PeerRunnerResultStatus = 'succeeded' | 'failed' | 'retried';
+
+/**
+ * Generic result type for all peer runners.
+ * Replaces the per-runner result interfaces (DreamerRunnerResult, etc.).
+ */
+export interface PeerRunnerResult<TOutput> {
+  readonly status: PeerRunnerResultStatus;
+  readonly taskId: string;
+  readonly runId?: string;
+  readonly artifactId?: string;
+  readonly resultRef?: string;
+  readonly contextHash?: string;
+  readonly output?: TOutput;
+  readonly errorCategory?: PDErrorCategory;
+  readonly failureReason?: string;
+  readonly attemptCount: number;
+}
+
+// ── Validation ───────────────────────────────────────────────────────────────
+
+/**
+ * Validation result returned by the abstract validateOutput() method.
+ * Replaces the per-runner validation result interfaces.
+ *
+ * Note: named PeerRunnerValidationResult to avoid collision with
+ * ValidationResult from rule-code-validator.ts.
+ */
+export interface PeerRunnerValidationResult {
+  readonly valid: boolean;
+  readonly errors: readonly string[];
+  readonly errorCategory?: PDErrorCategory;
+}
+
+// ── Internal context types (used by base class) ──────────────────────────────
+
+/** Context for the retryOrFail decision. */
+export interface FailureContext {
+  readonly taskId: string;
+  readonly task: TaskRecord;
+  readonly errorCategory: PDErrorCategory;
+  readonly failureReason: string;
+}
+
+/** Context for validation error handling. */
+export interface ValidationErrorContext {
+  readonly taskId: string;
+  readonly task: TaskRecord;
+  readonly errors: readonly string[];
+  readonly errorCategory?: PDErrorCategory;
+}
